chore: Move test utils to newsroom.tests for other repos
